Move ecma_policy.dart
