+++ pkgs/intl4x/pubspec.yaml
@@ -1,5 +1,5 @@
 name: intl4x
-description: An absolute bare-bones web app.
+description: A lightweight modular library for internationalization (i18n) functionality.
 version: 0.0.1-dev
 repository: https://github.com/dart-lang/i18n/tree/main/pkgs/intl4x
