+++ pkgs/intl4x/test/ecma/datetime_format_test.dart
@@ -5,7 +5,6 @@
 @TestOn('browser')
 library;
 
-import 'package:intl4x/ecma_policy.dart';
 import 'package:intl4x/intl.dart';
 import 'package:intl4x/src/datetime_format/datetime_format_options.dart';
 import 'package:test/test.dart';
